* moves UserCount component to presenter folder in users
