+++ src/app/users/list/component/component.html
@@ -3,6 +3,5 @@
     <h3 mat-subheader>List</h3>
     <mat-list-item class="usernames__element" *ngFor="let username of usernames">{{username}}</mat-list-item>
     <h3 mat-subheader>Count</h3>
-    <mat-list-item> <app-user-counter></app-user-counter></mat-list-item>
   </mat-list>
 </ng-container>
